Work on more graph checks
Graph checks now check for duplicate node or group definitions,
and also, unneeded edges are now removed so that we can perform
edge checks later on more easily.

The function `cleanup_nodes_vector` has been renamed to `cleanup_
unneeded_elements_in_vector`, because that's what it does, since
it already is a template. It's also applied to the edges vector
now which works because both have a property `needed` and the
vectors that store the pointers to them in the group need to be
treated in the same way anyway.